refactor: refactor platform-dependant code in a new module and fix error handling
