+++ src/store/scheme.rs
@@ -1,0 +1,6 @@
+static KNOWN_REMOTE_SCHEMES: &[&str] = &["http", "https", "registry"];
+
+/// Returns true if the scheme is a known remote scheme.
+pub(crate) fn is_known_remote_scheme(scheme: &str) -> bool {
+    KNOWN_REMOTE_SCHEMES.contains(&scheme)
+}
